move ViewOffset and ViewDataAnalysis to a separate header
